[amqp-heartbeats] reintroduce patches dir :(
