+++ lib/sensu/transport/rabbitmq.rb
@@ -3,22 +3,7 @@ gem "amqp", "1.5.0"
 require "amqp"
 
 require File.join(File.dirname(__FILE__), "base")
-
-module AMQP
-  class Session
-    def send_heartbeat
-      if tcp_connection_established? && !reconnecting?
-        send_frame(AMQ::Protocol::HeartbeatFrame)
-        if !@handling_skipped_heartbeats && @last_server_heartbeat
-          if @last_server_heartbeat < (Time.now - (self.heartbeat_interval * 2))
-            logger.error("detected missing amqp heartbeats")
-            self.handle_skipped_heartbeats
-          end
-        end
-      end
-    end
-  end
-end
+require File.join(File.dirname(__FILE__), "patches", "amqp")
 
 module Sensu
   module Transport
